improve test suite
As you can see from https://py.checkio.org/mission/date-and-time-converter/publications/leggewie/python-3/second-bool-for-plural-s-and-f-string/ the current test suite is incomplete (I made a mistake in how I use s1 and s2, they need to be used the other way round).  Only two out of the four possibilities for hour/hours and minute/minutes combinations are tested.  

I've changed one more test to make sure that "7 minutes" is not accepted as "07 minutes".